Actualizando cantidadDeRespuestasConDeterminadoBodys, test.

+++ src/test/kotlin/ar/edu/unahur/obj2/servidorWeb/repl.ws.kts
@@ -1,3 +1,3 @@
 package ar.edu.unahur.obj2.servidorWeb
 
-listOf(1, 8, 10).average().toInt()
+listOf(1, 8, 10).average()
